TEFile now supports huge files, standardising size_t to uint64_t for position; active update/installation now checks for the required file existence; anaglyph color stereo rendering is added; a bug in olxstr::SetLength is fixed; AFileSystem _DoesExist now takes an extra argument argument to force httpFS to check for the file existence;

+++ trunk/sdl/hashing.h
@@ -90,8 +90,8 @@ protected:
 
   olxcstr DoDigest(IInputStream& stream)  {
     unsigned char cbf[64];
-    const size_t blocks = stream.GetSize()/64;
-    for( size_t i=0; i < blocks; i++ )  {
+    const uint64_t blocks = stream.GetSize()/64;
+    for( uint64_t i=0; i < blocks; i++ )  {
       stream.Read(cbf, 64);
       Tools::hs_copy(cbf, bf, 64);
       Impl::digest64( bf );
